fix FP For used-before-assignment if terminating func is used
