+++ doc/whatsnew/fragments/7563.false_positive
@@ -1,0 +1,3 @@
+Fix ``used-before-assignment`` false positive when else branch calls ``sys.exit`` or similar terminating functions.
+
+Closes #7563
