+++ pylint/checkers/base/basic_checker.py
@@ -668,27 +668,6 @@ class BasicChecker(_BasicChecker):
             ):
                 self.add_message("misplaced-format-function", node=call_node)
 
-    @staticmethod
-    def _is_terminating_func(node: nodes.Call) -> bool:
-        """Detect call to exit(), quit(), os._exit(), or sys.exit()."""
-        if (
-            not isinstance(node.func, nodes.Attribute)
-            and not (isinstance(node.func, nodes.Name))
-            or isinstance(node.parent, nodes.Lambda)
-        ):
-            return False
-
-        qnames = {"_sitebuiltins.Quitter", "sys.exit", "posix._exit", "nt._exit"}
-
-        try:
-            for inferred in node.func.infer():
-                if hasattr(inferred, "qname") and inferred.qname() in qnames:
-                    return True
-        except (StopIteration, astroid.InferenceError):
-            pass
-
-        return False
-
     @utils.only_required_for_messages(
         "eval-used",
         "exec-used",
@@ -698,7 +677,7 @@ class BasicChecker(_BasicChecker):
     )
     def visit_call(self, node: nodes.Call) -> None:
         """Visit a Call node."""
-        if self._is_terminating_func(node):
+        if utils.is_terminating_func(node):
             self._check_unreachable(node, confidence=INFERENCE)
         self._check_misplaced_format_function(node)
         if isinstance(node.func, nodes.Name):
