+++ pylint/checkers/utils.py
@@ -2138,3 +2138,24 @@ def is_singleton_const(node: nodes.NodeNG) -> bool:
     return isinstance(node, nodes.Const) and any(
         node.value is value for value in SINGLETON_VALUES
     )
+
+
+def is_terminating_func(node: nodes.Call) -> bool:
+    """Detect call to exit(), quit(), os._exit(), or sys.exit()."""
+    if (
+        not isinstance(node.func, nodes.Attribute)
+        and not (isinstance(node.func, nodes.Name))
+        or isinstance(node.parent, nodes.Lambda)
+    ):
+        return False
+
+    qnames = {"_sitebuiltins.Quitter", "sys.exit", "posix._exit", "nt._exit"}
+
+    try:
+        for inferred in node.func.infer():
+            if hasattr(inferred, "qname") and inferred.qname() in qnames:
+                return True
+    except (StopIteration, astroid.InferenceError):
+        pass
+
+    return False
