+++ tests/functional/u/used/used_before_assignment_else_return.txt
@@ -1,4 +1,4 @@
-used-before-assignment:25:14:25:19:invalid:Using variable 'error' before assignment:CONTROL_FLOW
-used-before-assignment:38:14:38:19:invalid_2:Using variable 'error' before assignment:CONTROL_FLOW
-used-before-assignment:50:14:50:19:invalid_3:Using variable 'error' before assignment:CONTROL_FLOW
-used-before-assignment:60:14:60:19:invalid_4:Using variable 'error' before assignment:CONTROL_FLOW
+used-before-assignment:26:14:26:19:invalid:Using variable 'error' before assignment:CONTROL_FLOW
+used-before-assignment:39:14:39:19:invalid_2:Using variable 'error' before assignment:CONTROL_FLOW
+used-before-assignment:51:14:51:19:invalid_3:Using variable 'error' before assignment:CONTROL_FLOW
+used-before-assignment:61:14:61:19:invalid_4:Using variable 'error' before assignment:CONTROL_FLOW
